Refactor ChatGPTSmallToolbar and add it to the ChatGPTLogin
