+++ feature-login/src/main/res/values/strings.xml
@@ -15,5 +15,5 @@
      limitations under the License.
 -->
 <resources>
-  <string name="toast_logged_in">Successfully logged-in!</string>
+  <string name="top_bar_login">chatgpt-android login</string>
 </resources>
